requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.1
mysql-connector-python==8.3.0 